requirements.txt
openai~=1.40.3
python-dotenv~=1.0.1
langchain~=0.2.12
InstructorEmbedding~=1.0.1